Window resize / input pane workaround
Input pane is a special case, because we intentionally place it half off
the screen so the lower part of the border is hidden. This works fine
when calling newwin(), but wresize() and mvwin() refuse to work when a
window is partially offscreen.

We could:
a) implement a special case for input pane and make the border
frame only 2 lines high instead of 3,
b) remove the input pane border entirely, or
c) display the entire border.

I'm going with c) for now, but I'm not entirely happy with the way it
looks. Subject to further change.